moved inference.js to services.
